pyproject.toml
[tool.poetry]
name = "messages-api-bot"
version = "0.1.0"
description = ""
authors = ["Dddarknight <9801677@gmail.com>"]
readme = "README.md"
packages = [{include = "messages_api_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
djangorestframework = "^3.14.0"
python-dotenv = "^0.21.0"
drf-yasg = "^1.21.4"
djangorestframework-simplejwt = "^5.2.2"
requests = "^2.28.1"
python-telegram-bot = "^20.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
